Uses the ksuther xpc branch to support sandboxing
The `https://github.com/ksuther/Sparkle/tree/xpc` branch of the Sparkle framework adds support for an app that uses sandboxing.  The current Sparkle framework (as of 1.11) can not be used with a sandboxed app.

The ksuther branch is not the perfect solution, but is currently the best one, as discussed in https://github.com/sparkle-project/Sparkle/issues/363

Uses the 1.10 version: Sparkle-xpc-6a20fcb46d